+++ cmd/skpr/project/info/command.go
@@ -1,23 +1,23 @@
-package get
+package info
 
 import (
 	"github.com/spf13/cobra"
 
-	v1get "github.com/skpr/cli/internal/command/project/get"
+	v1get "github.com/skpr/cli/internal/command/project/info"
 )
 
 var (
 	cmdExample = `
-  # Get project details
-  skpr project get`
+  # Get project information
+  skpr project info`
 )
 
-// NewCommand creates a new cobra.Command for 'get' sub command
+// NewCommand creates a new cobra.Command for 'info' sub command
 func NewCommand() *cobra.Command {
 	command := v1get.Command{}
 
 	cmd := &cobra.Command{
-		Use:                   "get",
+		Use:                   "info",
 		Args:                  cobra.NoArgs,
 		DisableFlagsInUseLine: true,
 		Short:                 "Get full details for the current project.",
